pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "HidOverI2C"
version = "0.1.1"
description = "Python library for communicating with HID-over-I2C devices"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENCE" }
authors = [
	{ name = "Skyler Mansfield" },
]
keywords = ["hid", "i2c", "hid-over-i2c", "embedded", "hardware"]
classifiers = [
	"Development Status :: 3 - Alpha",
	"Intended Audience :: Developers",
	"License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)",
	"Operating System :: OS Independent",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3 :: Only",
	"Programming Language :: Python :: 3.9",
	"Programming Language :: Python :: 3.10",
	"Programming Language :: Python :: 3.11",
	"Programming Language :: Python :: 3.12",
	"Topic :: Software Development :: Libraries",
	"Topic :: System :: Hardware",
]
dependencies = [
	
]

[project.optional-dependencies]
SMBus2 = ["smbus2>=0.5.0"]
Examples = [
	"hidapi",
	"ft260",
	"hid-parser",
]


[project.urls]
Homepage = "https://github.com/Skyler84/pyHidOverI2C"
Repository = "https://github.com/Skyler84/pyHidOverI2C"
Issues = "https://github.com/Skyler84/pyHidOverI2C/issues"

[tool.hatch.build.targets.wheel]
packages = ["hidoveri2c"]

[tool.hatch.build.targets.sdist]
include = [
	"/hidoveri2c",
	"/README.md",
	"/LICENCE",
	"/pyproject.toml",
]
